Refactor: Flatten modules and remove redundant abstractions (Razor)
- Flatten `src/ast/` into `src/ast.rs`.
- Flatten `src/codegen/` into `src/codegen.rs`.
- Flatten `src/errors/` into `src/errors.rs` (merged `assembly.rs`, `messages.rs`).
- Flatten `src/parser/` into `src/parser.rs` (merged `grammar.rs`, `numerals.rs`, `recursion.rs`) and move `grammar.pest` to `src/parser.pest`.
- Merge `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs` to improve cohesion.
- Simplify imports and remove redundant module wrappers.
- Update tests and documentation to reflect new structure.

This reduces file count and folder depth, adhering to KISS and YAGNI principles.
